pyproject.toml
[tool.poetry]
name = "es"
version = "0.1.0"
description = ""
authors = ["dev <rvotaisrvota@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django = "^4.1.7"
docker-compose = "^1.29.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
